requirements.txt
certifi==2019.6.16
chardet==3.0.4
idna==2.8
oauthlib==3.0.1
PySocks==1.7.0
python-dotenv==0.10.3
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
tornado==6.0.2
tweepy==3.7.0
urllib3==1.25.3
